requests to the api are reduced
