+++ websites/leaderboard-new/service/components/data-view/data-update-know.js
@@ -1,21 +1,8 @@
-import React, { useState, useEffect } from "react";
-import api from "../../utils/api";
+import React from "react";
 import "../../styles/main.scss";
 
-const DataUpdateKnow = () => {
-  const [API_LastUpdated, SetAPI_LastUpdated] = useState(null);
-
-  useEffect(() => {
-    const fetchStats = async () => {
-      const apiStats = await api.getState();
-      if (apiStats.data) {
-        let CleanDate = apiStats.data.refreshed.slice(0, -5);
-        SetAPI_LastUpdated(CleanDate);
-      }
-    };
-    fetchStats();
-  }, []);
-
+const DataUpdateKnow = ({API_LastUpdated}) => {
+  
   return (
     <div className="box-info">
       <div>
