+++ websites/leaderboard-new/service/components/layout/left-side.js
@@ -1,32 +1,11 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import { useRouter } from "next/router";
 import Link from "next/link";
 import "../../styles/main.scss";
 import TweetBasodino from "../tweet-basodino";
-import api from "../../utils/api";
 
-const LeftSide = () => {
+const LeftSide = ({hash,copyCodeToClipboard}) => {
   const router = useRouter();
-  const [hash, setHash] = useState(
-    "16Uiu2HAm7KxaBkgd9ENvhf5qAkp1c6Q5Q1dXe8HBDzxLN4SxAVw6"
-  );
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const response = await api.getAllData();
-      if (response.data) setHash(response.data.address);
-    };
-    fetchData();
-  }, []);
-
-  const [modal, setModal] = useState(false);
-  const copyCodeToClipboard = () => {
-    navigator.clipboard.writeText(hash);
-    setModal(true);
-    setTimeout(() => {
-      setModal(false);
-    }, 4000);
-  };
 
   return (
     <section className="area-left-desktop">
