+++ websites/leaderboard-new/service/components/layout/right-side.js
@@ -4,14 +4,14 @@ import DataBoxCloud from "../data-view/data-box-cloud";
 import DataUpdateKnow from "../data-view/data-update-know";
 import "../../styles/main.scss";
 
-const RightSide = () => {
+const RightSide = ({ address, channel, API_LastUpdated }) => {
   return (
     <section className="right-side only-desktop-view">
       <CopieParagraph />
       <hr />
-      <DataBoxCloud />
+      <DataBoxCloud address={address} channel={channel} />
       <hr />
-      <DataUpdateKnow />
+      <DataUpdateKnow API_LastUpdated={API_LastUpdated} />
     </section>
   );
 };
