+++ websites/leaderboard-new/service/pages/index.js
@@ -5,8 +5,9 @@ import BoxDataTable from "../components/data-view/box-data-table";
 import SearchBar from "../components/micro-components/search-bar";
 import TrCustom from "../components/micro-components/tr-custom";
 import SuperBoxSearch from "../components/micro-components/super-box-search";
-
 import api from "../utils/api";
+
+
 
 export default function Home() {
   const columnsDefaults = [
@@ -64,13 +65,13 @@ export default function Home() {
   useEffect(() => {
     callAPI();
     return () =>{
-      setData(undefined)
+      setData(undefined);
     }
   }, []);
 
+
   useEffect(() => {
     let count = 0;
-
     if (nodes) {
       if (nodes.length) {
         count = nodes.length;
@@ -85,7 +86,7 @@ export default function Home() {
         }
       }
     }
-    setMatch(count);
+    setMatch(count); 
   }, [searchTerm]);
 
   const getIntBase = (key) => {
